Add GitHub sync webhook endpoint and logic to process GH issues webhooks
The received webhooks calls are validated against their signature using
the authors_github["webhook_key"] encrypted credential, which has to be
set up in GitHub as the webhook key.
The events processed are only the issue related actions "opened", "edited",
and "deleted"; and they accordingly create, update or delete authors in
the database.
The processing of the events is done asynchronously through an Active Job,
this allows both reprocessing errors, and scaling the event processing
separately from the API backend.